Fix language switcher when used with geoip
If GeoIP is enabled but there is no translation of the node for the
redirected language, some other translation will be rendered without
changing the path, so the language switcher can’t identify which node
it is displayed on (no link matches the current path).
It still needs to render an “active” link as parent item.