Clientside validation for input-groups
Treat input groups as 1 element and display all error messages after.
The event callbacks need to look out for the error message matching the
validated input. Clientside validation always adds a `for` attribute,
even if the error messages aren’t labels, so we can safely use it for
identifying the error message.